Extract search bar into a component
